LayoutTests/imported/w3c/web-platform-tests/workers/interfaces/WorkerGlobalScope/location/redirect-sharedworker-expected.txt

CONSOLE MESSAGE: line 8: ReferenceError: Can't find variable: SharedWorker

FAIL WorkerLocation with redirects: classic shared workers ReferenceError: Can't find variable: SharedWorker
